More kernel tweaks (#6428)
* kernel: drop unused return value of UseTmpPerm

* kernel: use BOOL in more places

* kernel: init buffer with memset

* kernel: use size_t in some places

* kernel: explain certain uses of 'extern inline'

* kernel: update some comments

+++ src/scanner.c
@@ -71,7 +71,7 @@ static void SyntaxErrorOrWarning(ScannerState * s,
         // error is reported, so the pretty-printed context would leak
         // a literal 0xFF byte to *errout* with no caret to anchor it.
         const char * line = GetInputLineBuffer(s->input);
-        const UInt len = strlen(line);
+        const size_t len = strlen(line);
         const int line_is_sentinel = (len == 0 || line[0] == '\377');
         if (!line_is_sentinel) {
             if (line[len-1] != '\n')
